refactor proxy setup
