+++ cmd/main.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/Vyary/otel-rev-proxy/internal/proxy"
-	"github.com/Vyary/otel-rev-proxy/internal/telemetry"
+	"github.com/Vyary/otel-rev-proxy/pkg/telemetry"
 	"go.opentelemetry.io/contrib/bridges/otelslog"
 )
 
@@ -18,12 +18,13 @@ var logger = otelslog.NewLogger("reverse-proxy")
 func main() {
 	if err := run(); err != nil {
 		logger.Error("failed to start the reverse proxy", "error", err)
-		slog.Error("failed to start the reverse proxy", "error", err)
 		os.Exit(1)
 	}
 }
 
 func run() error {
+	port := os.Getenv("PORT")
+
 	ctx, stop := signal.NotifyContext(context.Background(), os.Interrupt, syscall.SIGTERM)
 	defer stop()
 
@@ -33,14 +34,14 @@ func run() error {
 	}
 	defer otelShutdown(context.Background())
 
-	srv, err := proxy.New()
+	srv, err := proxy.New(port)
 	if err != nil {
 		return err
 	}
 
 	srvErr := make(chan error, 1)
 	go func() {
-		slog.Info("Starting Reverse Proxy...")
+		slog.Info("Starting the Reverse Proxy...")
 		srvErr <- srv.ListenAndServe()
 	}()
 
@@ -54,7 +55,7 @@ func run() error {
 	ctxTO, cancel := context.WithTimeout(context.Background(), 20*time.Second)
 	defer cancel()
 
-	slog.Info("Shutting Down Reverse Proxy...")
+	slog.Info("Shutting Down the Reverse Proxy...")
 
 	return srv.Shutdown(ctxTO)
 }
